All image and external files used
The images are what make up the UI in general. They are systematically named so that the code can pull it from its directory into the correct object. The CSV file is where a record of the runs is kept and modified. 3 example saves are inside the file as a proof of concept.