Feature classes for use in main file.
The folder contains every tool needed in classes, methods, and variables to help guide a system to completion. The files are imported within each other as they benefit from the functionality each one provides. The interconnected structure means a change in a variable has an immediate and complementary response. Nothing is hard coded so changing the value of a single variable in the "variables.py" file adjusts every relevant code associated with it. For example, changing the image's width causes the cards to realign, fonts to get bigger so that they are readable, and messages to shift to an appropriate position.