Main files.
The files contain the while loop to create the turns and all classes and methods that were provided before. The entire process is relatively smooth in its logical flow and is clearly labeled. Any changes in the class files they benefit from will change the result but cannot change the nature of the game such as turn orders and the nature of the betting mechanism. Just like in more professional settings, the files do not have any advanced code that should be done by other classes. The "main_with_history" file has a couple of changes but ultimately is an upgrade of the "main" file.